Set up build to use Paketo build packs for Spring Native.
Move dependency to the native JAR into the common build as we have code to compile to tweak the native image. That code needs to be compilable in non-native scenarios as well.

Use buildpack based native image builds. Remove offending libraries: both Spring Boot DevTools and the Thin Launcher are not native-ready yet. Added NativeConfiguration to capture a few manual hints still needed.

Declare Money as the type being returned from MonetaryAmountAttributeConverter.convertToEntityAttribute(…) so that it is automatically added to the reflection configuration (makes spring-projects-experimental/spring-native#826 obsolete).